% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit_sentence.R
\name{credit_sentence}
\alias{credit_sentence}
\title{Credit Sentence}
\usage{
credit_sentence(data, group_var)
}
\arguments{
\item{data}{A dataframe.}

\item{group_var}{A grouping variable.}
}
\value{
A dataframe with two columns: credit_statement and sentence.
}
\description{
This function takes a dataframe and a grouping variable as inputs. It then
groups the dataframe by the grouping variable column using group_by(),
and then calculates the credit_statement and sentence columns using
summarise(). In the paste0() function for the sentence column,
it wraps the name variable in \if{html}{\out{<b>}} tags to make it bold.
}
\examples{

df_credit <- read.csv(
  "https://raw.githubusercontent.com/Global-Health-Engineering/manuscript-qechairquality/main/data/credit.csv") |>
   tidyr::unite("name", first_name:surname, sep = "")

credit_sentence(data = df_credit, group_var = name)

}
