% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_summarize.R
\name{group_summarize}
\alias{group_summarize}
\title{Group and summarize data by multiple variables}
\usage{
group_summarize(data, group_vars, summary_vars)
}
\arguments{
\item{data}{A data frame}

\item{group_vars}{A character vector of variable names to group by}

\item{summary_vars}{A character vector of variable names to calculate summary statistics for}
}
\value{
A data frame with one row for each combination of grouping variables and one column for each summary statistic
}
\description{
This function groups a data frame by multiple variables and calculates summary statistics for multiple variables.
}
\examples{

# Load mtcars dataset

data(mtcars)

# Group and summarize data by number of cylinders and transmission type
group_summarize(mtcars, c("cyl", "am"), c("hp", "mpg"))

}
